[DOP-20962] Use test celery instance, add separate transfer fixture
