+++ docker-compose.test.yml
@@ -78,6 +78,7 @@ services:
       context: .
       target: test
     command: --loglevel=info -Q test_queue
+    entrypoint: [python, -m, celery, -A, tests.test_integration.celery_test, worker, --max-tasks-per-child=1]
     env_file: .env.docker
     volumes:
       - ./syncmaster:/app/syncmaster
@@ -90,7 +91,7 @@ services:
         condition: service_healthy
       rabbitmq:
         condition: service_healthy
-    profiles: [worker, s3, oracle, hdfs, hive, all]
+    profiles: [worker, scheduler, s3, oracle, hdfs, hive, all]
 
   test-postgres:
     image: postgres
