+++ tests/test_integration/test_scheduler/scheduler_fixtures/mocker_fixtures.py
@@ -2,14 +2,14 @@ import asyncio
 
 import pytest
 from apscheduler.triggers.cron import CronTrigger
-from pytest_mock import MockerFixture
+from pytest_mock import MockerFixture, MockType
 
 from syncmaster.scheduler.transfer_job_manager import TransferJobManager
 from syncmaster.worker.config import celery
 
 
 @pytest.fixture
-def mock_send_task_to_tick(mocker: MockerFixture):
+def mock_send_task_to_tick(mocker: MockerFixture) -> MockType:
     original_to_thread = asyncio.to_thread
     return mocker.patch(
         "asyncio.to_thread",
@@ -18,7 +18,7 @@ def mock_send_task_to_tick(mocker: MockerFixture):
 
 
 @pytest.fixture
-def mock_add_job(mocker: MockerFixture, transfer_job_manager: TransferJobManager):
+def mock_add_job(mocker: MockerFixture, transfer_job_manager: TransferJobManager) -> MockType:
     original_add_job = transfer_job_manager.scheduler.add_job
     return mocker.patch.object(
         transfer_job_manager.scheduler,
